small refactoring per discussion with Marco
